requirements.txt
spacy>=3.7.5
google-generativeai>=0.8.0
fastapi>=0.100.0
uvicorn>=0.20.0
requests>=2.31.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
en_core_web_trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.7.3/en_core_web_trf-3.7.3-py3-none-any.whl
